+++ CHANGES.rst
@@ -11,6 +11,7 @@ Version 0.8.0 (UNRELEASED)
 - Changes ``list`` to display interactive session status.
 - Changes ``list`` command to stop including workflow progress and workspace size by default. Adds
   ``--include-progress`` and ``--include-workspace-size`` options to add them to the output.
+- Changes Yadage workflow specification loading to be done in ``reana-commons``.
 
 Version 0.7.5 (UNRELEASED)
 --------------------------